config: module compression supported but don't autocompress
We're leaving module compression enabled, with Zstd selected, but this
commit disables CONFIG_MODULE_COMPRESS_ALL.  I want my kernel
installers (and installed initrd) to be as small as possible (sigh, as
less big at this point), and initramfs compression is much more
effective if the modules aren't all compressed individually.  We're
trading about 7M of installed space in /lib/modules for a 5M reduction
in installer size.

265M	FOO.modzstd
272M	FOO.modzstdnoauto
137M	kernel-installer-6.12.25-mdl+.sh.modzstd
132M	kernel-installer-6.12.25-mdl+.sh.modzstdnoauto